# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management Service API",
        "description": "A high-performance Laravel-based API service for managing translations across multiple locales with advanced tagging and search capabilities.",
        "contact": {
            "email": "admin@translationservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Translation Management Service API Server"
        }
    ],
    "paths": {
...
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get paginated list of translations",
                "operationId": "ecee4c59cba20fd72cdc6b9c1d8bc084",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter by locale code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "mobile"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in keys and values",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "welcome"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by exact key",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "welcome.message"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Translation"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations for frontend use",
                "operationId": "6b6344befbdfe96832018f642d51ce02",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale code to export",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Filter by tag names",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Key-value pairs of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "welcome.message": "Welcome to our application",
                                    "nav.home": "Home",
                                    "btn.save": "Save"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Locale": {
                "title": "Locale",
                "description": "Locale model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "en"
                    },
                    "name": {
                        "type": "string",
                        "example": "English"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "title": "Tag",
                "description": "Tag model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "mobile"
                    },
                    "description": {
                        "type": "string",
                        "example": "Mobile application translations"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Translation": {
                "title": "Translation",
                "description": "Translation model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "key": {
                        "type": "string",
                        "example": "welcome.message"
                    },
                    "value": {
                        "type": "string",
                        "example": "Welcome to our application"
                    },
                    "locale_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "locale": {
                        "$ref": "#/components/schemas/Locale"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token for authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}